6301-workspace reducer and action rework: - separeted organization and material/help actions from workspace actions file - organization workspace is separated to its own reducer file - corresponding changes to files to fix imports because previously mentioned changes
