+++ muikku-core-plugins/src/main/resources/META-INF/resources/scripts/src/actions/index.ts
@@ -137,27 +137,8 @@ import {
   UPDATE_WORKSPACES_ALL_PROPS,
   UPDATE_WORKSPACES_ACTIVE_FILTERS,
   UPDATE_WORKSPACE,
-  UPDATE_WORKSPACES_SET_CURRENT_MATERIALS,
-  UPDATE_WORKSPACES_SET_CURRENT_MATERIALS_ACTIVE_NODE_ID,
-  UPDATE_WORKSPACES_SET_CURRENT_MATERIALS_REPLIES,
-  UPDATE_ORGANIZATION_WORKSPACES_AVAILABLE_FILTERS_EDUCATION_TYPES,
-  UPDATE_ORGANIZATION_WORKSPACES_AVAILABLE_FILTERS_STATE_TYPES,
-  UPDATE_ORGANIZATION_WORKSPACES_AVAILABLE_FILTERS_CURRICULUMS,
-  UPDATE_ORGANIZATION_WORKSPACES_ALL_PROPS,
-  UPDATE_ORGANIZATION_SELECTED_WORKSPACE,
-  UPDATE_ORGANIZATION_WORKSPACES_ACTIVE_FILTERS,
-  UPDATE_ORGANIZATION_WORKSPACES_STATE,
   UPDATE_CURRENT_COMPOSITE_REPLIES_UPDATE_OR_CREATE_COMPOSITE_REPLY_STATE_VIA_ID_NO_ANSWER,
-  UPDATE_MATERIAL_CONTENT_NODE,
-  DELETE_MATERIAL_CONTENT_NODE,
-  INSERT_MATERIAL_CONTENT_NODE,
-  UPDATE_PATH_FROM_MATERIAL_CONTENT_NODES,
   UPDATE_WORKSPACES_EDIT_MODE_STATE,
-  UPDATE_WORKSPACES_SET_CURRENT_HELP,
-  UPDATE_ORGANIZATION_TEMPLATES,
-  UPDATE_WORKSPACES_AVAILABLE_FILTERS_ORGANIZATIONS,
-  UPDATE_ORGANIZATION_SELECTED_WORKSPACE_STAFF_SELECT_STATE,
-  UPDATE_ORGANIZATION_SELECTED_WORKSPACE_STUDENT_SELECT_STATE,
   UPDATE_CURRENT_WORKSPACE_ACTIVITY,
   UPDATE_CURRENT_WORKSPACE_ASESSMENT_REQUESTS,
   UPDATE_AVAILABLE_CURRICULUMS,
@@ -299,6 +280,29 @@ import {
 // Guidance events are no longer being
 import { UPDATE_CALENDAR_EVENTS_STATUS } from "./main-function/calendar";
 import { SET_ACTIVE_EASY_TO_USE_TOOL } from "./easy-to-use-functions";
+import {
+  UPDATE_WORKSPACES_SET_CURRENT_MATERIALS,
+  UPDATE_WORKSPACES_SET_CURRENT_MATERIALS_ACTIVE_NODE_ID,
+  UPDATE_WORKSPACES_SET_CURRENT_MATERIALS_REPLIES,
+  UPDATE_MATERIAL_CONTENT_NODE,
+  DELETE_MATERIAL_CONTENT_NODE,
+  INSERT_MATERIAL_CONTENT_NODE,
+  UPDATE_PATH_FROM_MATERIAL_CONTENT_NODES,
+  UPDATE_WORKSPACES_SET_CURRENT_HELP,
+} from "./workspaces/material";
+import {
+  UPDATE_ORGANIZATION_WORKSPACES_AVAILABLE_FILTERS_EDUCATION_TYPES,
+  UPDATE_ORGANIZATION_WORKSPACES_AVAILABLE_FILTERS_STATE_TYPES,
+  UPDATE_ORGANIZATION_WORKSPACES_AVAILABLE_FILTERS_CURRICULUMS,
+  UPDATE_ORGANIZATION_WORKSPACES_ALL_PROPS,
+  UPDATE_ORGANIZATION_WORKSPACES_ACTIVE_FILTERS,
+  UPDATE_ORGANIZATION_WORKSPACES_STATE,
+  UPDATE_ORGANIZATION_SELECTED_WORKSPACE,
+  UPDATE_ORGANIZATION_SELECTED_WORKSPACE_STAFF_SELECT_STATE,
+  UPDATE_ORGANIZATION_SELECTED_WORKSPACE_STUDENT_SELECT_STATE,
+  UPDATE_ORGANIZATION_TEMPLATES,
+  UPDATE_WORKSPACES_AVAILABLE_FILTERS_ORGANIZATIONS,
+} from "./workspaces/organization";
 import {
   JOURNALS_COMMENTS_CREATE,
   JOURNALS_COMMENTS_DELETE,
