+++ muikku-core-plugins/src/main/resources/META-INF/resources/scripts/src/actions/workspaces/helpers.ts
@@ -355,3 +355,24 @@ export async function loadCurrentWorkspaceJournalsHelper(
     });
   }
 }
+
+/**
+ * reuseExistantValue
+ * @param conditional conditional
+ * @param existantValue existantValue
+ * @param otherwise otherwise
+ */
+export function reuseExistantValue(
+  conditional: boolean,
+  existantValue: any,
+  otherwise: () => any
+) {
+  if (!conditional) {
+    return null;
+  }
+  if (existantValue) {
+    return existantValue;
+  }
+
+  return otherwise();
+}
