+++ muikku-core-plugins/src/main/resources/META-INF/resources/scripts/src/components/base/material-editor/confirm-remove-page-with-linked-answers-dialog.tsx
@@ -10,10 +10,10 @@ import { bindActionCreators } from "redux";
 import { WorkspaceMaterialEditorType } from "~/reducers/workspaces";
 import {
   setWorkspaceMaterialEditorState,
-  SetWorkspaceMaterialEditorStateTriggerType,
   updateWorkspaceMaterialContentNode,
+  SetWorkspaceMaterialEditorStateTriggerType,
   UpdateWorkspaceMaterialContentNodeTriggerType,
-} from "~/actions/workspaces";
+} from "~/actions/workspaces/material";
 
 /**
  * ConfirmPublishRemovePageWithLinkedAnswersDialogProps
