+++ muikku-core-plugins/src/main/resources/META-INF/resources/scripts/src/components/base/material-editor/delete-dialog.tsx
@@ -12,9 +12,9 @@ import {
   WorkspaceMaterialEditorType,
 } from "~/reducers/workspaces";
 import {
-  deleteWorkspaceMaterialContentNode,
   DeleteWorkspaceMaterialContentNodeTriggerType,
-} from "~/actions/workspaces";
+  deleteWorkspaceMaterialContentNode,
+} from "~/actions/workspaces/material";
 
 /**
  * DeleteWorkspaceMaterialDialogProps
